fixed bearing vectors, added eigen, fixed color equ for LUT

+++ include/ollieRosTools/PreProc.hpp
@@ -37,7 +37,7 @@ private:
 
     // Smoothing variables
     int smootherNr;
-    double sigmaX, sigmaY;
+    float sigmaX, sigmaY;
     cv::Size kSize;
     int k;
